Comments Query rename: - comments-template -> comment-template - comments-query -> comments-query-loop
