+++ packages/block-library/src/comment-template/block.json
@@ -1,9 +1,9 @@
 {
 	"apiVersion": 2,
-	"name": "core/comments-template",
-	"title": "Comments Template",
+	"name": "core/comment-template",
+	"title": "Comment Template",
 	"category": "design",
-	"parent": [ "core/comments-query" ],
+	"parent": [ "core/comments-query-loop" ],
 	"description": "Contains the block elements used to render a comment, like the title, date, author, avatar and more.",
 	"textdomain": "default",
 	"usesContext": [
