+++ packages/block-library/src/comment-template/save.js
@@ -3,6 +3,6 @@
  */
 import { InnerBlocks } from '@wordpress/block-editor';
 
-export default function CommentsTemplateSave() {
+export default function CommentTemplateSave() {
 	return <InnerBlocks.Content />;
 }
