+++ packages/block-library/src/comments-query-loop/block.json
@@ -1,6 +1,6 @@
 {
 	"apiVersion": 2,
-	"name": "core/comments-query",
+	"name": "core/comments-query-loop",
 	"title": "Comments Query Loop",
 	"category": "theme",
 	"description": "An advanced block that allows displaying post comments based on different query parameters and visual configurations.",
