+++ test/integration/fixtures/blocks/core__comment-template.html
@@ -1,0 +1,1 @@
+<!-- wp:comment-template /-->
