+++ test/integration/fixtures/blocks/core__comment-template.json
@@ -1,0 +1,10 @@
+[
+	{
+		"clientId": "_clientId_0",
+		"name": "core/comment-template",
+		"isValid": true,
+		"attributes": {},
+		"innerBlocks": [],
+		"originalContent": ""
+	}
+]
