+++ test/integration/fixtures/blocks/core__comment-template.parsed.json
@@ -1,0 +1,9 @@
+[
+	{
+		"blockName": "core/comment-template",
+		"attrs": {},
+		"innerBlocks": [],
+		"innerHTML": "",
+		"innerContent": []
+	}
+]
