+++ test/integration/fixtures/blocks/core__comments-query-loop.json
@@ -1,0 +1,14 @@
+[
+	{
+		"clientId": "_clientId_0",
+		"name": "core/comments-query-loop",
+		"isValid": true,
+		"attributes": {
+			"queryPerPage": 4,
+			"queryOffset": 0,
+			"tagName": "div"
+		},
+		"innerBlocks": [],
+		"originalContent": "<div class=\"wp-block-comments-query-loop\"></div>"
+	}
+]
