+++ test/integration/fixtures/blocks/core__comments-query-loop.parsed.json
@@ -1,0 +1,11 @@
+[
+	{
+		"blockName": "core/comments-query-loop",
+		"attrs": {},
+		"innerBlocks": [],
+		"innerHTML": "\n<div class=\"wp-block-comments-query-loop\"></div>\n",
+		"innerContent": [
+			"\n<div class=\"wp-block-comments-query-loop\"></div>\n"
+		]
+	}
+]
