+++ test/integration/fixtures/blocks/core__comments-query-loop.serialized.html
@@ -1,0 +1,3 @@
+<!-- wp:comments-query-loop -->
+<div class="wp-block-comments-query-loop"></div>
+<!-- /wp:comments-query-loop -->
